Migrate component-related logic to dedicated task module; update utils
