+++ tasks/index.js
@@ -1,0 +1,5 @@
+const { generate: generateComponent } = require('./component');
+
+module.exports = {
+  generateComponent,
+};
